Files required for determining AIS establishment
Please note that this script will do a pull of occurrence records for species listed in the Aquatic Invasive Species Regulations plus the unlisted species (see unlisted_AIS.csv) and, therefore, the data produced will be current to whatever date the data pull is conducted. Data used in our manuscript only included species records up to Jan 2022. Some species such as marbled crayfish and red swamp crayfish, which are now reported within the assessment area, were not known to occur in Nova Scotia at the time of the assessment. If you would like the data for either species then please add those names to the unlisted_AIS.csv file. A csv file of all the AIS present (not including the two aforementioned crayfish) is provided here (AIS_data.csv) and is current as of the date of this post.